Optimized player movement
I have optimized the player movement, both the keyboard and touch. It is now so there is one pull target that is used for the player is moving towards with a speed. And this pull target is moved on arrow clicks or swipes. Instead of the old one where there where 3 pull targets and they where moved forward when moving forward. This new code is much much more logical and this is actually the player movement I wanted from the beginning because now the player can move diagonal, because there is only one pull target and the player is always being pulled/moving towards this LaneTarget.