Rotation to player
Added rotation to player, so now when the player is changing lane, the player rotates a little bit, so it looks like he is driving in that direction.
It was a bit annoying to make, because he was rotating sideways, and i could not get a pivot point to work. So I went the ver simple way, by make a picture of timmy sideways, and using that as the player sprite, this was, he is going straight with the rotation. The player is never seen sideways, as I have rotated the gameobject from the beginning. Probably the wrong way, but I couldn't figure out the math to calculate the rotation other ways, and the pivot point would'nt do as I wanted.