ObstacleSpawner + small changes
I made the ObstacleSpawner, so far it has 6 different obstacles, and some of them have er 75% chance to spawn an extra obstacle a little behind the original in a random lane. So you might become trapped in a lane because there is a bus in the middle and an extra in the lane your in. But when you see the extra it is too late. So you can get hit by the bus or the Blue car. This is both good and bad, good because right now the game is too easy, but with this feature it makes it a bit harder. But it is very rare to see this obstacle.