Background music and Mute button
I have added background music. It is the Gold for cash song, and I tried to loop it as good as I could.
I have also added a mute button in the top right corner because people might get annoyed quickly, I will have to figure out how to add the mute setting in a constant variable so you don't have to mute it every time the scene restarts or you start the game.