Consistent mute setting
Now the setting the player has preferred last time playing is the chosen one when opening the game next time. So if the player had muted it completly, it will be completly muted next time playing. This is just so it wont be annoying you every time you open the game, but mostly because it will choose this if you die and restart the gamescene, the game will still be muted.