Score works now
I have added a score text element. that displays the score.
The score is calculated in the PlayerStats scripts, in the update method it adds a ScoreIncreasev variable and times that with Time.deltaTime, this is then Rounded and made a string so it won't show the decimals, only whole numbers.