Loose screen and tab to start
I have added a loose screen, so when you loose, a screen pops up with your score and the current highscore. You can then press start. Small changes needs to be fixed, for example the highscore is about 10-20 points to high on the end screen, this is easily fixed by using Invoke somewhere, just need to figure out where I set the highscore and call the EndGame method. I also need to fix the Obstaclespawner, because on restart it calls the WhichObstacle method, and if there are a method that calls an obstacle in the start, then sometimes 4 obstacles are called at the same time, which makes it impossible to avoid them.
Other then that, I have made the Tab To start animation that plays on the gamescene. So when you start the game or press restart, you have to tab on the screen to start the game.