Crash sound, plus fixes
I have fixed the issue with the shown highscore was 10-20 points to high.
I have also fixed the issue where too many obstacles where spawned when you clicked the restart button.
On the Start method it check if the Time is lower than 1, if it is, the game has just been opened and an obstacle needs to be spawned. It is not, the Update will call the method that spawns an obstacle, because Time is larger than lastTimeSpawned + spawnEverySec.

I have also added a crash sound effect for when you are hit by a car. It might not be the right one, but so far it works fine. I can easily change the sounds as it is just variables in the audioManager and easy to switch to another.