Fixed motor mute on restart + Implemented tiers
I have fixed an issue where the motor was unmuted on restart, no matter if it was muted before. I fixed that, so it will be muted if it where muted before.

I have also implemented Tiers, so every time the score increases with 200, the tier increases which means the speed will increase on everything. So it will be become harder. I thought it was a bit to easy to get a high score, so to make the games a bit faster I made this.