ALOT of timmy sounds
I have added A LOT of Timmy sound effects.
One of which will be the greeting sound when he is close to Jimmy on the sidewalk. If he greets Jimmy succesfully, the player will be granted more points.

I might need to make a coin System, so the player feel rewarded when he gets a high score. So he will be able to Pimp Timmy's ride, maybe the cars and even the street he is driving. So the player will keep playing my game.

With this, I can make Ingame purchases, so the player will be able to buy the ingame coins and thus be able to buy skins for the different stuff.

This might first be when I have released the game in Beta. I will start working on the freemium aspect of the game.