Fixed obstacles on top of each other + gap in background
I have fixed the obstacles that spawned on top of each other. If an obstacle spawns on top of another, it gets moved a little up on the y-axis, so they are no longer on top of each other. This may be a wrong approach, because if a cop car drives in the rear of another car on screen, it will be moved a little up as well. So I might need to do a little check if they are within screen, if so, the car in front needs to speed up instead of making car behind back up.

There was also a small gap in the background every other time. This is fixed by putting a static background behind the two moving, so the gap is not noticable.