Button check
Added Button check for tapping. If the player taps the muteBtn in the top right corner before starting the game, the game will not start, but it will change the Audio Settings.
(Cannot test this on mobile, Unity Remote won't work on my PC for some reason. So i will test tomorrow on Laptop)