Fixed button click before start
I fixed the CheckButtonClick method. It would not recognize the button as a gameobject. So it still started the game when clicking the mutebutton.  I made it so it checks the the position of the tap instead, so if the tap position is in the top right corner, it do not start the game and only clicks the button. If it not in the corner, it starts the game.