Boost + Explosion particles
I have now made the Particle system for the boost. So when Timmy is boosting, there is fire shooting out the back.
I have also made the explosion effect for the cars. So if Timmy is boosting, and drives into a car, the car will blow up, and it will give the player the same amount of points as it would have retracted normally.