PickupSpawner script
I have made the Pickup spawner, it spawns a pickup item every X seconds. And it can never be the same one twice in a row.
Also made a quick change in the PlayerStats, there is now a isProtected, isBoosting and isShielded. Because isProtected were buggy if you both had a shield and where boosting. The is protected could be set to false when it should be true.